cleanup, refactor bases
