+++ Makefile
@@ -3,7 +3,7 @@
 dist:
 	python setup.py sdist
 
-checks:
+check:
 	pipenv run flake8
 	pipenv run mypy .
 	pipenv run bento check
