+++ jetkit/db/__init__.py
@@ -1,4 +1,6 @@
-from jetkit.db.query import BaseQuery
+from jetkit.db.bases import SQLA
 from jetkit.db.model import BaseModel
+from jetkit.db.query import BaseQuery
 
-__all__ = ("BaseQuery", "BaseModel")
+
+__all__ = ("BaseQuery", "BaseModel", "SQLA")
